+++ lib/kitchen/directions/bake_screenreader_spans.rb
@@ -7,13 +7,19 @@ module Kitchen
       # Additional screenreader spans can be added below.
       def self.v1(book:)
         book.search('u[data-effect="double-underline"]').each do |element|
-          element.add_previous_sibling(
+          element.prepend(child:
             '<span data-screenreader-only="true">double underline</span>'
+          )
+          element.append(child:
+            '<span data-screenreader-only="true">end double underline</span>'
           )
         end
         book.search('u[data-effect="underline"]').each do |element|
-          element.add_previous_sibling(
+          element.prepend(child:
             '<span data-screenreader-only="true">underline</span>'
+          )
+          element.append(child:
+            '<span data-screenreader-only="true">end underline</span>'
           )
         end
       end